requirements.txt
faker>=15.0.0
pymongo>=4.0.0
python-dotenv>=1.0.0
requests>=2.25.0
flask>=2.0.0
dnspython>=2.0.0